microservices/visual-generation-service/requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
google-cloud-storage==2.14.0
google-cloud-vision==3.4.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.23.3
python-json-logger==2.0.7
Pillow==10.0.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
stability-sdk==0.8.3  # For Stability AI integration
replicate==0.11.0  # For additional AI models 